requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
djangorestframework-simplejwt==5.2.0
django-extensions==3.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
gunicorn==20.1.0
whitenoise==6.2.0
dj-database-url==1.0.0
django-cors-headers==3.13.0
cryptography==38.0.3
django-filter==23.2
